backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dateutil
passlib
bcrypt
pyjwt
